requirements.txt
streamlit==1.29.0
openai==1.3.0
requests==2.31.0
python-dotenv==1.0.0
json5==0.9.14
typing-extensions==4.8.0